+++ src/thorin/be/c/hls.cpp
@@ -1,0 +1,9 @@
+#include "hls.h"
+
+namespace thorin::c_be {
+
+HLSCodeGen::HLSCodeGen(World& world, const Cont2Config& kernel_config, int opt, bool debug)
+    : CodeGen(world, kernel_config, Lang::HLS, debug)
+{}
+
+}
